weather update conf added

+++ RESTServices/Weather/Coord.cs
@@ -3,10 +3,10 @@ namespace Husky_sTestBot.RESTServices.Weather{
 
     public class Coord    {
         [JsonPropertyName("lon")]
-        public int Lon { get; set; } 
+        public double Lon { get; set; } 
 
         [JsonPropertyName("lat")]
-        public int Lat { get; set; } 
+        public double Lat { get; set; } 
     }
 
 }
